Execute a command

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -3,4 +3,5 @@
 exec{'kill_killmenow_process' :
   path    => '/bin/',
   command => 'pkill killmenow',
+  unless  => 'pgrep killmenow'
 }
